+++ matlab/stairparam.m
@@ -15,14 +15,14 @@ con_options = struct('Algorithm', 'sqp'); % 'OutputFcn', @outfun,,'PlotFcns',@op
 
 %disp(v_r);
 
-    subplot(2, 1, hs);
-    plot(xi, zi, 'x');
-    axis equal tight
-    hold on;
-    plot(xi, z_r, 'o')
-    axis equal tight
-    hold off;
-    xlabel(hs);
+    % subplot(2, 1, hs);
+    % plot(xi, zi, 'x');
+    % axis equal tight
+    % hold on;
+    % plot(xi, z_r, 'o')
+    % axis equal tight
+    % hold off;
+    % xlabel(hs);
 
 %% Delta function - calculates the difference between real-z and template-z
     function [se] = delta(v)