+++ matlab/test.m
@@ -2,8 +2,8 @@ function [] = test(xi,zi,h)
 % x = linspace(1,1,10);
 % plot(x);
 % disp(dx);
-subplot(2,1,h)
-plot(xi,zi,'s');
-axis equal tight
+% subplot(2,1,h)
+% plot(xi,zi,'s');
+% axis equal tight
 
 end